Reorganize scorecard components into their own directory + fix spacing  (#170)
Reorganize scorecard components into their own directory + fix spacing
