+++ client/src/views/ScorecardView.vue
@@ -25,12 +25,10 @@
             <SidePanel />
           </div>
         </div>
-
         <div class='column is-hidden-mobile'>
-          <NationwideMap height='60vh' :enableBasicMap='true' />
+          <NationwideMap height='70vh' :enableBasicMap='true' />
         </div>
       </div>
-
       <div class='section has-text-centered' v-if='showResults'>
         <div class='h1-header-large'>
           {{ ScorecardMessages.TAKE_ACTION_HEADER }}
@@ -52,9 +50,7 @@
             @onButtonClick='copyToClipboard' />
         </div>
       </div>
-
       <ContactCitySection class='section' v-if='showLslrSection' :city='city' />
-
       <ScorecardSummaryPanel v-if='showResults' />
       <ActionSection
         class='section'
@@ -74,27 +70,27 @@
 </template>
 
 <script lang='ts'>
-import PredictionPanel from '../components/PredictionPanel.vue';
-import ActionSection from '../components/ActionSection.vue';
+import PredictionPanel from '../components/scorecard_page/PredictionPanel.vue';
+import ActionSection from '../components/scorecard_page/ActionSection.vue';
 import { defineComponent } from 'vue';
 import { router, SCORECARD_BASE } from '../router';
-import ScorecardSummaryPanel from '../components/ScorecardSummaryPanel.vue';
+import ScorecardSummaryPanel from '../components/scorecard_page/ScorecardSummaryPanel.vue';
 import { ScorecardMessages } from '../assets/messages/scorecard_messages';
 import { Titles } from '../assets/messages/common';
 import NationwideMap from '../components/NationwideMap.vue';
-import LslrSection from '../components/LslrSection.vue';
+import LslrSection from '../components/scorecard_page/LslrSection.vue';
 import { dispatch, useSelector } from '../model/store';
 import { LeadDataState } from '../model/states/lead_data_state';
 import { GeoDataState } from '../model/states/geo_data_state';
-import ScorecardMapSearchBar from '../components/ScorecardMapSearchBar.vue';
-import SidePanel from '../components/SidePanel.vue';
+import ScorecardMapSearchBar from '../components/scorecard_page/ScorecardMapSearchBar.vue';
+import SidePanel from '../components/scorecard_page/ScorecardSidePanel.vue';
 import { City } from '../model/states/model/geo_data';
 import { Status } from '../model/states/status_state';
 import { getParcel, getWaterSystem } from '../model/slices/lead_data_slice';
 import { GeoDataUtil } from '../util/geo_data_util';
 import Loading from 'vue-loading-overlay';
 import 'vue-loading-overlay/dist/vue-loading.css';
-import ContactCitySection from '../components/ContactCitySection.vue';
+import ContactCitySection from '../components/scorecard_page/ContactCitySection.vue';
 
 /**
  * Container for SearchBar and MapContainer.
